requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.2
seaborn==0.12.1
nltk==3.7
openai==0.27.8
PyYAML==6.0
tqdm==4.64.1